Introduce TypeIgnore Feature
with the IgnoreType attribute provides the possiblity
to skip types defined by a glob or regex pattern.

Additionally some rules where added to the rule engine
with the same idea. With Black/ Whitelist rules types
may be excluded in the rule engines. This makes it
posssible to document the exluded type in the output.
In opposite the IgnoreType simple skips the type at
directly in the analyzer.